Normalization/pyproject.toml
[project]
name = "xLSTF"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "lightning>=2.5.0.post0",
    "pandas>=2.2.3",
    "prettytable>=3.12.0",
    "scikit-learn>=1.6.1",
    "torch>=2.5.1",
    "xlstm>=2.0.2",
]
